moved get_response_gene() from gene_to_panels.py to panel_app_api_functions.py
